--- desafiostone/src/main/java/br/com/stone/desafiostone/DesafiostoneApplication.java
@@ -1,14 +1,0 @@
-package br.com.stone.desafiostone;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DesafiostoneApplication {
-
-	
-	public static void main(String[] args) {
-		SpringApplication.run(DesafiostoneApplication.class, args);
-	}
-
-}
