--- desafiostone/src/main/java/br/com/stone/desafiostone/entity/Response.java
@@ -1,15 +1,0 @@
-package br.com.stone.desafiostone.entity;
-
-import org.springframework.stereotype.Component;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Component
-@Getter
-@Setter
-public class Response {
-
-  private String mensagem;
-
-}
