--- desafiostone/src/main/java/br/com/stone/desafiostone/entity/Transaction.java
@@ -1,28 +1,0 @@
-package br.com.stone.desafiostone.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToOne;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Table(name = "tb02_transaction")
-@Getter
-@Setter
-public class Transaction {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    private String client_id;
-    private String client_name;
-    private int total_to_pay;
-
-    @OneToOne
-    private CreditCard credit_card;
-  
-}
