--- desafiostone/src/main/java/br/com/stone/desafiostone/repository/CreditCardRepository.java
@@ -1,9 +1,0 @@
-package br.com.stone.desafiostone.repository;
-
-import org.springframework.data.repository.CrudRepository;
-
-import br.com.stone.desafiostone.entity.CreditCard;
-
-public interface CreditCardRepository extends CrudRepository<CreditCard, Long>{
-  
-}
