--- desafiostone/src/main/java/br/com/stone/desafiostone/repository/TransactionRepository.java
@@ -1,9 +1,0 @@
-package br.com.stone.desafiostone.repository;
-
-import org.springframework.data.repository.CrudRepository;
-
-import br.com.stone.desafiostone.entity.Transaction;
-
-public interface TransactionRepository extends CrudRepository<Transaction , Long>{
-  
-}
