--- desafiostone/src/main/resources/application.properties
@@ -1,4 +1,0 @@
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://localhost:3306/desafiostone
-spring.datasource.username=root
-spring.datasource.password=Juli@2233
