--- desafiostone/src/test/java/br/com/stone/desafiostone/DesafiostoneApplicationTests.java
@@ -1,13 +1,0 @@
-package br.com.stone.desafiostone;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DesafiostoneApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
