+++ lojastartwars/src/main/java/br/com/stone/desafiostone/DesafiostoneApplication.java
@@ -1,0 +1,15 @@
+package br.com.stone.lojastartwars;
+
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DesafiostoneApplication {
+
+	
+	public static void main(String[] args) {
+		SpringApplication.run(DesafiostoneApplication.class, args);
+	}
+
+}
