+++ lojastartwars/src/main/java/br/com/stone/desafiostone/controller/HistoryController.java
@@ -1,0 +1,46 @@
+package br.com.stone.lojastartwars.controller;
+
+import java.util.Optional;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import br.com.stone.lojastartwars.entity.History;
+import br.com.stone.lojastartwars.entity.Response;
+import br.com.stone.lojastartwars.service.HistoryService;
+
+@RestController
+@RequestMapping("/starstore/history")
+@CrossOrigin(origins = "*")
+public class HistoryController {
+
+    @Autowired
+    HistoryService HistoryService;
+
+    @DeleteMapping("/remover/{codigo}")
+    public ResponseEntity<Response> remover(@PathVariable long codigo) {
+        return HistoryService.remover(codigo);
+    }
+
+    @PostMapping("/cadastrar")
+    public ResponseEntity<?> cadastrar(@RequestBody History history) {
+        return HistoryService.cadastrarAlterar(history, "cadastrar");
+    }
+
+    @PutMapping("/alterar")
+    public ResponseEntity<?> alterar(@RequestBody History history) {
+        return HistoryService.cadastrarAlterar(history, "alterar");
+    }
+
+    @GetMapping("/listar")
+    public Iterable<History> listar() {
+        return HistoryService.listar();
+    }
+
+    @GetMapping("/listar/{clientId}")
+    public ResponseEntity<Optional<History>> listarPorCliente(@PathVariable long clientId) {
+        Optional<History> History = HistoryService.encontrarPorId(clientId);
+        return ResponseEntity.ok(History);
+    }
+}
