+++ lojastartwars/src/main/java/br/com/stone/desafiostone/entity/CreditCard.java
@@ -1,0 +1,26 @@
+package br.com.stone.lojastartwars.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Table(name = "tb03_creditcard")
+@Getter
+@Setter
+public class CreditCard {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+    private String card_number;
+    private double value;
+    private int cvv;
+    private String card_holder_name;
+    private String exp_date;
+  
+}
