+++ lojastartwars/src/main/java/br/com/stone/desafiostone/entity/Product.java
@@ -1,0 +1,27 @@
+package br.com.stone.lojastartwars.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Table(name = "tb01_products")
+@Getter
+@Setter
+public class Product {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private long id;
+  private String title;
+  private int price;
+  private String zipcode;
+  private String seller;
+  private String thumbnailHd;
+  private String date;
+
+}
