+++ lojastartwars/src/main/java/br/com/stone/desafiostone/entity/Response.java
@@ -1,0 +1,15 @@
+package br.com.stone.lojastartwars.entity;
+
+import org.springframework.stereotype.Component;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Component
+@Getter
+@Setter
+public class Response {
+
+  private String mensagem;
+
+}
