+++ lojastartwars/src/main/java/br/com/stone/desafiostone/repository/CreditCardRepository.java
@@ -1,0 +1,9 @@
+package br.com.stone.lojastartwars.repository;
+
+import org.springframework.data.repository.CrudRepository;
+
+import br.com.stone.lojastartwars.entity.CreditCard;
+
+public interface CreditCardRepository extends CrudRepository<CreditCard, Long>{
+  
+}
