+++ lojastartwars/src/main/java/br/com/stone/desafiostone/repository/ProductRepository.java
@@ -1,0 +1,9 @@
+package br.com.stone.lojastartwars.repository;
+
+import org.springframework.data.repository.CrudRepository;
+
+import br.com.stone.lojastartwars.entity.Product;
+
+public interface ProductRepository extends CrudRepository <Product, Long> {
+  
+}
