+++ lojastartwars/src/main/java/br/com/stone/desafiostone/repository/TransactionRepository.java
@@ -1,0 +1,9 @@
+package br.com.stone.lojastartwars.repository;
+
+import org.springframework.data.repository.CrudRepository;
+
+import br.com.stone.lojastartwars.entity.Transaction;
+
+public interface TransactionRepository extends CrudRepository<Transaction , Long>{
+  
+}
