+++ lojastartwars/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+spring.jpa.hibernate.ddl-auto=update
+spring.datasource.url=jdbc:mysql://localhost:3306/loja_startwars
+spring.datasource.username=root
+spring.datasource.password=Juli@2233
