+++ lojastartwars/src/test/java/br/com/stone/lojastartwars/DesafiostoneApplicationTests.java
@@ -1,0 +1,13 @@
+package br.com.stone.lojastartwars;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DesafiostoneApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
